rproxy-backend/requirements.txt
flask==1.1.2
Werkzeug==1.0.1
marshmallow==3.7.1
requests==2.24.0
flask-sqlalchemy==2.4.4
psycopg2==2.8.5
alembic==1.4.2
celery==4.4.7
PyCryptodome==3.9.8
gunicorn==20.0.4
gevent==20.6.2